chore: remove ConfirmSwapModalMod and move logic to index
